Auto-correct desired replicas on missing webhook_job completion event
While testing #1179, I discovered that ARC sometimes stop resyncing RunnerReplicaSet when the desired replicas is greater than the actual number of runner pods.
This seems to happen when ARC missed receiving a workflow_job completion event but it has no way to decide if it is either (1) something went wrong on ARC or (2) a loadbalancer in the middle or GitHub or anything not ARC went wrong. It needs a standard to decide it.

So I added a hard-coded 30 seconds timeout(can be made customizable later) to prevent runner pod recreation.